- Moved Unit Tests into their own Fixture as per GitHub Discussion. - Renamed Tests to conform to existing standards
